requirements.txt
praat-parselmouth
praat-textgrids
requests
TTS
